cinema tickets last version

+++ src/repositories/actor.reponsitory.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { ActorDto } from 'src/modules/actor/dto/actor.dto';
@@ -10,42 +14,41 @@ export class ActorReponsitory {
     @InjectModel(Actor.name) private readonly actorModel: Model<Actor>,
   ) {}
   async createActor(actorDto: ActorDto): Promise<any> {
-    const existsActor = await this.actorModel.findOne({
-      name: actorDto.name,
-    });
-    if (existsActor) {
-      return 'Actor already exists';
+    try {
+      const existsActor = await this.actorModel.findOne({
+        name: actorDto.name,
+      });
+      if (existsActor) throw new ConflictException('Actor already exists');
+      const createActor = new this.actorModel(actorDto);
+      if (!createActor) throw new UnauthorizedException('Create Fail');
+      createActor.save();
+      return createActor;
+    } catch (error) {
+      return error.message;
     }
-    const createActor = new this.actorModel(actorDto);
-    if (!createActor) {
-      return {
-        status: 'error',
-        message: 'Create Failed',
-      };
-    }
-    return {
-      data: createActor.save(),
-      message: 'Successfully',
-    };
   }
 
-  async getAllactor(): Promise<any> {
+  async getAllactor(page: number): Promise<any> {
+    const pageSize = 10;
+    const skip = (page - 1) * pageSize;
+    const getAll = await this.actorModel.find({}).skip(skip).limit(pageSize);
+    return getAll;
+  }
+
+  async getAllactorNoPage(): Promise<any> {
     const getAll = await this.actorModel.find({});
     return getAll;
   }
+
   async getactor(actorId: any): Promise<any> {
     const getActor = await this.actorModel.findById(actorId);
     if (!getActor) {
-      return {
-        status: 'error',
-        message: 'Get Failed',
-      };
+      return 'Get Failed';
     }
-    return {
-      data: getActor,
-      message: 'Successfully',
-    };
+    return getActor;
   }
+
+  
 
   async updateActor(actorId: any, dataUpdate: any): Promise<any> {
     const update = await this.actorModel.findByIdAndUpdate(
@@ -56,24 +59,19 @@ export class ActorReponsitory {
       },
     );
     if (!update) {
-      return {
-        status: 'error',
-        message: 'Update Failed',
-      };
+      return 'Update Failed';
     }
     return update;
-     
   }
 
-  async deleteActor(actorId: any): Promise<any> {
+  async deleteActor(actorId: any, password: any): Promise<any> {
+    if (password != '8888') {
+      return 'You do not have sufficient authority to delete';
+    }
     const deleteactor = await this.actorModel.findByIdAndDelete(actorId);
     if (!deleteactor) {
-      return {
-        status: 'error',
-        message: 'Delete Failed',
-      };
+      return 'Delete Failed';
     }
     return deleteactor;
-    
   }
 }
