+++ src/repositories/cinema.reponsitory.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CinemaDto } from 'src/modules/cinema/dto/cinema.dto';
@@ -14,40 +18,27 @@ export class CinemaReponsitory {
       const existCinema = await this.cinemaModel.findOne({
         name: cinemaDto.name,
       });
-      if (existCinema) {
-        return 'Cinema already exists';
-      }
+      if (existCinema) throw new ConflictException('Cinema already exists');
       const createcinema = new this.cinemaModel(cinemaDto);
-      if (!createcinema) {
-        return {
-          status: 'error',
-          message: 'Create Failed',
-        };
-      }
-      return {
-        data: createcinema.save(),
-        message: 'Successfully',
-      };
+      if (!createcinema) throw new UnauthorizedException('Create Fail');
+      createcinema.save();
+      return createcinema;
     } catch (error) {
-      return error.response;
+      return error.message;
     }
   }
 
   async getAllCinema(): Promise<any> {
+  
     const getAll = await this.cinemaModel.find({});
     return getAll;
   }
   async getCinema(cinemaId: any): Promise<any> {
     const getcinema = await this.cinemaModel.findById(cinemaId);
     if (!getcinema) {
-      return {
-        status: 'error',
-        message: 'Get Failed',
-      };
+      return 'Get Failed';
     }
-    return  getcinema;
-
-  
+    return getcinema;
   }
 
   async updateCinema(cinemaId: any, dataUpdate: any): Promise<any> {
@@ -59,26 +50,21 @@ export class CinemaReponsitory {
       },
     );
     if (!update) {
-      return {
-        status: 'error',
-        message: 'Update Failed',
-      };
+      return 'Update Failed';
     }
-    return  update;
 
-   
+    return update;
   }
 
-  async deleteCinema(cinemaId: any): Promise<any> {
+  async deleteCinema(cinemaId: any, password: any): Promise<any> {
+    if (password != '8888') {
+      return 'You do not have sufficient authority to delete';
+    }
     const deletecinema = await this.cinemaModel.findByIdAndDelete(cinemaId);
     if (!deletecinema) {
-      return {
-        status: 'error',
-        message: 'Delete Failed',
-      };
+      return 'Delete Failed';
     }
-    return deletecinema;
 
-    
+    return deletecinema;
   }
 }
