+++ src/repositories/director.reponsitory.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { DirectorDto } from 'src/modules/director/dto/director.dto';
@@ -11,37 +15,38 @@ export class DirectorsReponsitory {
   ) {}
 
   async createDirectorDto(directorDto: DirectorDto): Promise<any> {
-    const existsDirector = await this.directorModel.findOne({
-      name: directorDto.name,
-    });
-    if (existsDirector) {
-      return 'Director already exists';
-    }
+    try {
+      const existsDirector = await this.directorModel.findOne({
+        name: directorDto.name,
+      });
+      if (existsDirector)
+        throw new ConflictException('Director already exists');
 
-    const createDirector = new this.directorModel(directorDto);
-    if (!createDirector) {
-      return {
-        status: 'error',
-        message: 'Create Failed',
-      };
+      const createDirector = new this.directorModel(directorDto);
+      if (!createDirector) throw new UnauthorizedException('Create Fail');
+      createDirector.save();
+      return createDirector;
+    } catch (error) {
+      error.message;
     }
-    return {
-      data: createDirector.save(),
-      message: 'Successfully',
-    };
   }
 
-  async getAlldirector(): Promise<any> {
+  async getAlldirector(page: number): Promise<any> {
+    const pageSize = 10;
+    const skip = (page - 1) * pageSize;
+    const getAll = await this.directorModel.find({}).skip(skip).limit(pageSize);
+    return getAll;
+  }
+
+  async getAlldirectorNoPage(): Promise<any> {
     const getAll = await this.directorModel.find({});
     return getAll;
   }
+
   async getdirector(directorId: any): Promise<any> {
     const getDirector = await this.directorModel.findById(directorId);
     if (!getDirector) {
-      return {
-        status: 'error',
-        message: 'Get Failed',
-      };
+      return 'Get Failed';
     }
     return getDirector;
   }
@@ -55,22 +60,19 @@ export class DirectorsReponsitory {
       },
     );
     if (!update) {
-      return {
-        status: 'error',
-        message: 'Update Failed',
-      };
+      return 'Update Failed';
     }
     return update;
   }
 
-  async deleteDirector(directorId: any): Promise<any> {
+  async deleteDirector(directorId: any, password: any): Promise<any> {
+    if (password != '8888') {
+      return 'You do not have sufficient authority to delete';
+    }
     const deletedirector =
       await this.directorModel.findByIdAndDelete(directorId);
     if (!deletedirector) {
-      return {
-        status: 'error',
-        message: 'Delete Failed',
-      };
+      return 'Delete Failed';
     }
     return deletedirector;
   }
