+++ src/repositories/movie.reponsitory.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { MovieDto } from 'src/modules/movie/dto/movie.dto';
@@ -8,32 +12,49 @@ import { Movie } from 'src/schemas/movie.schema';
 export class MovieReponsitory {
   constructor(@InjectModel(Movie.name) readonly movieModel: Model<Movie>) {}
   async create(movieDto: MovieDto): Promise<any> {
-    const existsMovie = await this.movieModel.findOne({ name: movieDto.name });
-    if (existsMovie) {
-      return 'Movie already exists';
-    }
-    const dataCreate = {
-      ...movieDto,
-      release_date: new Date(movieDto.release_date),
-    };
-    const create = new this.movieModel(dataCreate);
-    if (!create) {
-      return {
-        status: 'error',
-        message: 'Create Movie Failed',
+    try {
+      const existsMovie = await this.movieModel.findOne({
+        name: movieDto.name,
+      });
+      if (existsMovie) throw new ConflictException('Movie already exists');
+
+      const dataCreate = {
+        ...movieDto,
+        release_date: new Date(movieDto.release_date),
       };
+      const create = new this.movieModel(dataCreate);
+      if (!create) throw new UnauthorizedException('Create Fail');
+      create.save();
+      return create;
+    } catch (error) {
+      return error.message;
     }
-    return {
-      data: create.save(),
-      message: 'Successfully',
-    };
   }
-  async getAllmovie(): Promise<any> {
-    const getAll = await this.movieModel.find({}).populate([
-      { path: 'genre', select: 'name image' },
-      { path: 'director', select: 'name image' },
-      { path: 'actor', select: 'name image' },
-    ]);
+  async getAllmovieForAdmin(page: number): Promise<any> {
+    const pageSize = 10;
+    const skip = (page - 1) * pageSize;
+    const getAll = await this.movieModel
+      .find({})
+      .populate([
+        { path: 'genre', select: 'name image' },
+        { path: 'director', select: 'name image' },
+        { path: 'actor', select: 'name image' },
+      ])
+      .sort({ release_date: -1 })
+      .skip(skip)
+      .limit(pageSize);
+    return getAll;
+  }
+  async getAllmovieForUser(): Promise<any> {
+    const getAll = await this.movieModel
+      .find({ release_status: { $ne: 'nc' } })
+      .sort({ release_date: -1 })
+      .populate([
+        { path: 'genre', select: 'name image' },
+        { path: 'director', select: 'name image' },
+        { path: 'actor', select: 'name image' },
+      ]);
+
     return getAll;
   }
   async getmovie(movieId: any): Promise<any> {
@@ -43,10 +64,7 @@ export class MovieReponsitory {
       { path: 'actor', select: 'name image' },
     ]);
     if (!getMovie) {
-      return {
-        status: 'error',
-        message: 'Get Movie Failed',
-      };
+      return 'Get Movie Failed';
     }
     return getMovie;
   }
@@ -54,6 +72,7 @@ export class MovieReponsitory {
   async getMovieByDirector(directorId: any): Promise<any> {
     const getMovieByDirector = await this.movieModel
       .find({
+        release_status: { $ne: 'nc' },
         director: directorId,
       })
       .populate([
@@ -62,25 +81,21 @@ export class MovieReponsitory {
         { path: 'actor', select: 'name image' },
       ]);
     if (!getMovieByDirector) {
-      return {
-        status: 'error',
-        message: 'Get Movie By Director Failed',
-      };
+      return 'Get Movie By Director Failed';
     }
     return getMovieByDirector;
   }
 
   async getMovieByActor(actorId: any): Promise<any> {
-    const getMovie = await this.movieModel.find({ actor: actorId }).populate([
-      { path: 'genre', select: 'name image' },
-      { path: 'director', select: 'name image' },
-      { path: 'actor', select: 'name image' },
-    ]);
+    const getMovie = await this.movieModel
+      .find({ actor: actorId, release_status: { $ne: 'nc' } })
+      .populate([
+        { path: 'genre', select: 'name image' },
+        { path: 'director', select: 'name image' },
+        { path: 'actor', select: 'name image' },
+      ]);
     if (!getMovie) {
-      return {
-        status: 'error',
-        message: 'Get Movie By Actor Failed',
-      };
+      return 'Get Movie By Actor Failed';
     }
     return getMovie;
   }
@@ -90,6 +105,7 @@ export class MovieReponsitory {
       const date = new Date(Date.now());
       const getMoviedc = await this.movieModel.find({
         release_date: { $lt: date },
+        release_status: { $ne: 'nc' },
       });
 
       if (getMoviedc.length > 0) {
@@ -124,14 +140,54 @@ export class MovieReponsitory {
     return regexResults;
   }
 
-  async getMovieByGenre(genre: any): Promise<any> {
-    const getMovie = await this.movieModel.find({ genre: genre });
+  async getMovieByGenre(genreId: any): Promise<any> {
+    const getMovie = await this.movieModel.find({
+      genre: genreId,
+      release_status: { $ne: 'nc' },
+    });
     return getMovie;
+  }
+
+  async updateMovieStopShow(movieId: any): Promise<any> {
+    try {
+      const movie = await this.movieModel.findById(movieId);
+      if (movie.release_status === 'nc') {
+        const update = await this.movieModel.findByIdAndUpdate(
+          movieId,
+          {
+            release_status: 'sc',
+          },
+          {
+            new: true,
+          },
+        );
+
+        if (!update) {
+          return 'Update Failed';
+        }
+      } else {
+        const update = await this.movieModel.findByIdAndUpdate(
+          movieId,
+          {
+            release_status: 'nc',
+          },
+          {
+            new: true,
+          },
+        );
+
+        if (!update) {
+          return 'Update Failed';
+        }
+        return update;
+      }
+    } catch (error) {
+      console.log('error', error.response);
+    }
   }
 
   async update(movieId: any, dataUpdate: any): Promise<any> {
     try {
-    
       const update = await this.movieModel.findByIdAndUpdate(
         movieId,
         dataUpdate,
@@ -139,29 +195,24 @@ export class MovieReponsitory {
           new: true,
         },
       );
-     
 
       if (!update) {
-        return {
-          status: 'error',
-          message: 'Update Failed',
-        };
+        return 'Update Failed';
       }
-     
+
       return update;
     } catch (error) {
-      console.log('error',error.response)
+      console.log('error', error.response);
     }
-  
   }
 
-  async delete(movieId: any): Promise<any> {
+  async delete(movieId: any, password: any): Promise<any> {
+    if (password != '8888') {
+      return 'You do not have sufficient authority to delete';
+    }
     const deletemovie = await this.movieModel.findByIdAndDelete(movieId);
     if (!deletemovie) {
-      return {
-        status: 'error',
-        message: 'Delete Failed',
-      };
+      return 'Delete Failed';
     }
     return deletemovie;
   }
